+++ CHANGELOG.md
@@ -105,6 +105,7 @@ UIIN-3437.
 * Condtionally evaluate DOM refs before calling `.focus()`. Fixes UIIN-3603.
 * Add `stripes-core.settings.read` permission to app and settings permissions. Refs UIIN-3544.
 * Add missing permission for number generator. Fixes UIIN-3609.
+* Inventory Holdings: Add number generator button "Generate call number" to additional holdings call numbers. Refs UIIN-3606.
 
 ## [13.0.13](https://github.com/folio-org/ui-inventory/tree/v13.0.13) (2026-03-11)
 [Full Changelog](https://github.com/folio-org/ui-inventory/compare/v13.0.12...v13.0.13)